Fix Edge not being removed from dom on delete
The existing code comments about a race condition for the node deletion.
I think the same thing is happening for the edge, although just wrapping
it in a timeout has proven to not be enough, so rather brutally, I've
implemented a retry loop which seems to fix it 99% of the time.